Update TypeService and TypeCOntroller for exception

+++ src/main/java/com/begginers/sim/inventory/service/TypeService.java
@@ -1,13 +1,13 @@
 package com.begginers.sim.inventory.service;
 
+import com.begginers.sim.inventory.exception.TypeNotFoundException;
 import com.begginers.sim.inventory.model.Type;
 import com.begginers.sim.inventory.repository.TypeRepository;
-import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -19,19 +19,23 @@ public class TypeService {
 
     private final TypeRepository typeRepository;
 
-    public List<Type> getAllTypes() {
-        return typeRepository.findAll();
-    }
-
-    public Optional<Type> getTypeById(Long id) {
-        return typeRepository.findById(id);
-    }
-
     public Type saveType(Type type) {
         return typeRepository.save(type);
     }
 
-    public void deleteType(Long id) {
+    public List<Type> getAllTypes() {
+        return typeRepository.findAll();
+    }
+
+    public Type getTypeById(Long id) throws TypeNotFoundException {
+        return typeRepository.findById(id)
+                .orElseThrow(() -> new TypeNotFoundException("Type not found with ID: " + id));
+    }
+
+    public void deleteType(Long id) throws TypeNotFoundException {
+        if (!typeRepository.existsById(id)) {
+            throw new TypeNotFoundException("Type not found with id: " + id);
+        }
         typeRepository.deleteById(id);
     }
 }
